Move DOM-independent zodiac functions into separate file
So it can be unit tested
