+++ source/Zodiac_compatibility/script.js
@@ -1,5 +1,10 @@
-import { mod } from "../utils.js";
-import { zodiacOrder, zodiacDateRanges, romantic } from "./data.js";
+import {
+  determineDateRangeLeft,
+  determineDateRangeRight,
+  determinePairing,
+  roundAngle,
+  textGenerator,
+} from "./zodiac-angles.js";
 
 // Get the wheel element
 //const wheel = document.querySelector('.wheel');
@@ -45,22 +50,6 @@ function rotateWheel2(event) {
   dateInput.style.background = "transparent";
   // Prevent the default scrolling behavior
   event.preventDefault();
-}
-
-function determineDateRangeLeft(angle) {
-  return zodiacDateRanges[getMappingLeft(angle)] ?? "unknown";
-}
-
-function determineDateRangeRight(angle) {
-  return zodiacDateRanges[getMappingRight(angle)] ?? "unknown";
-}
-
-// Add the event listener for the mouse wheel event
-//wheel.addEventListener('wheel', rotateWheel);
-// Function to round the angle to the nearest multiple of 30
-function roundAngle(angle) {
-  const rounded = Math.round(angle / 30) * 30;
-  return mod(rounded, 360);
 }
 
 // Function to handle the mouseout event
@@ -135,49 +124,3 @@ button.addEventListener("click", () => {
     pairing_text.innerHTML = textGenerator(pair[0], pair[1]);
   }, 2);
 });
-
-function getMappingLeft(angle) {
-  angle = mod(angle, 360);
-  for (const [i, zodiac] of zodiacOrder.entries()) {
-    if (angle < (i + 1) * 60) {
-      return zodiac;
-    }
-  }
-  return "unknown";
-}
-function getMappingRight(angle) {
-  angle = mod(angle, 360);
-  if (angle < 180) {
-    for (const [i, zodiac] of zodiacOrder.slice(6).entries()) {
-      if (angle < (i + 1) * 60) {
-        return zodiac;
-      }
-    }
-  } else {
-    for (const [i, zodiac] of zodiacOrder.slice(0, 6).entries()) {
-      if (angle - 180 < (i + 1) * 60) {
-        return zodiac;
-      }
-    }
-  }
-  return "unknown";
-}
-
-function determinePairing(angle1, angle2) {
-  angle1 = mod(angle1, 360);
-  angle2 = mod(angle2, 360);
-  const angle1Mapping = getMappingLeft(angle1);
-  const angle2Mapping = getMappingRight(angle2);
-  return [angle1Mapping, angle2Mapping];
-}
-
-// generate results text
-
-//text generator
-function textGenerator(one, two) {
-  return (
-    romantic.get(`${one} and ${two}`) ??
-    romantic.get(`${two} and ${one}`) ??
-    "An error has occurred"
-  );
-}
